chore: move example to test folder
