+++ test/example-async.js
@@ -132,15 +132,4 @@ function build (opts) {
   return fastify
 }
 
-if (require.main === module) {
-  const fastify = build({
-    logger: {
-      level: 'info'
-    }
-  })
-  fastify.listen({ port: 3000, host: '0.0.0.0' }, err => {
-    if (err) throw err
-  })
-}
-
 module.exports = build
